p-zalo/requirements.txt
robotframework
robotframework-Selenium2Library
schedule
robotframework-requests
selenium
webdriver-manager